# Fix timing issues