--- a/SDL/Kernels.h
+++ b/SDL/Kernels.h
@@ -12,8 +12,9 @@
 
 namespace SDL {
   ALPAKA_FN_ACC ALPAKA_FN_INLINE void rmQuintupletFromMemory(struct SDL::quintuplets& quintupletsInGPU,
-                                                             unsigned int quintupletIndex) {
-    quintupletsInGPU.isDup[quintupletIndex] = true;
+                                                             unsigned int quintupletIndex,
+                                                             bool secondpass = false) {
+    quintupletsInGPU.isDup[quintupletIndex] |= 1 + secondpass;
   };
 
   ALPAKA_FN_ACC ALPAKA_FN_INLINE void rmPixelTripletFromMemory(struct SDL::pixelTriplets& pixelTripletsInGPU,
@@ -212,10 +213,8 @@ namespace SDL {
           float phi1 = __H2F(quintupletsInGPU.phi[ix]);
           float score_rphisum1 = __H2F(quintupletsInGPU.score_rphisum[ix]);
 
-          for (unsigned int jx1 = globalThreadIdx[2]; jx1 < nQuintuplets_lowmod; jx1 += gridThreadExtent[2]) {
+          for (unsigned int jx1 = globalThreadIdx[2] + ix1 + 1; jx1 < nQuintuplets_lowmod; jx1 += gridThreadExtent[2]) {
             unsigned int jx = quintupletModuleIndices_lowmod + jx1;
-            if (ix == jx)
-              continue;
 
             float eta2 = __H2F(quintupletsInGPU.eta[jx]);
             float phi2 = __H2F(quintupletsInGPU.phi[jx]);
@@ -231,10 +230,9 @@ namespace SDL {
 
             int nMatched = checkHitsT5(ix, jx, quintupletsInGPU);
             if (nMatched >= 7) {
-              if (score_rphisum1 > score_rphisum2) {
+              if (score_rphisum1 >= score_rphisum2) {
                 rmQuintupletFromMemory(quintupletsInGPU, ix);
-              } else if ((score_rphisum1 == score_rphisum2) && (ix < jx)) {
-                rmQuintupletFromMemory(quintupletsInGPU, ix);
+                break;
               }
             }
           }
@@ -260,7 +258,7 @@ namespace SDL {
 
         unsigned int quintupletModuleIndices_lowmod1 = rangesInGPU.quintupletModuleIndices[lowmod1];
 
-        for (unsigned int lowmodIdx2 = globalThreadIdx[2]; lowmodIdx2 < *(rangesInGPU.nEligibleT5Modules);
+        for (unsigned int lowmodIdx2 = globalThreadIdx[2] + lowmodIdx1; lowmodIdx2 < *(rangesInGPU.nEligibleT5Modules);
              lowmodIdx2 += gridThreadExtent[2]) {
           uint16_t lowmod2 = rangesInGPU.indicesOfEligibleT5Modules[lowmodIdx2];
           unsigned int nQuintuplets_lowmod2 = quintupletsInGPU.nQuintuplets[lowmod2];
@@ -271,7 +269,7 @@ namespace SDL {
 
           for (unsigned int ix1 = 0; ix1 < nQuintuplets_lowmod1; ix1 += 1) {
             unsigned int ix = quintupletModuleIndices_lowmod1 + ix1;
-            if (quintupletsInGPU.partOfPT5[ix])
+            if (quintupletsInGPU.partOfPT5[ix] || (quintupletsInGPU.isDup[ix] & 1))
               continue;
 
             for (unsigned int jx1 = 0; jx1 < nQuintuplets_lowmod2; jx1++) {
@@ -279,7 +277,7 @@ namespace SDL {
               if (ix == jx)
                 continue;
 
-              if (quintupletsInGPU.partOfPT5[jx])
+              if (quintupletsInGPU.partOfPT5[jx] || (quintupletsInGPU.isDup[jx] & 1))
                 continue;
 
               float eta1 = __H2F(quintupletsInGPU.eta[ix]);
@@ -303,12 +301,11 @@ namespace SDL {
               int nMatched = checkHitsT5(ix, jx, quintupletsInGPU);
               if (dR2 < 0.001f || nMatched >= 5) {
                 if (score_rphisum1 > score_rphisum2) {
-                  rmQuintupletFromMemory(quintupletsInGPU, ix);
-                  continue;
-                }
-                if ((score_rphisum1 == score_rphisum2) && (ix < jx)) {
-                  rmQuintupletFromMemory(quintupletsInGPU, ix);
-                  continue;
+                  rmQuintupletFromMemory(quintupletsInGPU, ix, true);
+                } else if (score_rphisum1 < score_rphisum2) {
+                  rmQuintupletFromMemory(quintupletsInGPU, jx, true);
+                } else {
+                  rmQuintupletFromMemory(quintupletsInGPU, (ix < jx ? ix : jx), true);
                 }
               }
             }