--- a/bin/sdl.cc
+++ b/bin/sdl.cc
@@ -298,12 +298,15 @@ int main(int argc, char** argv)
 void run_sdl()
 {
     SDL::Dev devAcc = alpaka::getDevByIdx(ALPAKA_ACCELERATOR_NAMESPACE::Platform{}, 0u);
-    SDL::QueueAcc queue(devAcc);
+    std::vector<SDL::QueueAcc> queues;
+    for (int s = 0; s < ana.streams; s++) {
+        queues.push_back(SDL::QueueAcc(devAcc));
+    }
 
     // Load various maps used in the SDL reconstruction
     TStopwatch full_timer;
     full_timer.Start();
-    loadMaps(devAcc, queue);
+    loadMaps(devAcc, queues[0]);
     float timeForMapLoading = full_timer.RealTime()*1000;
 
     if (ana.do_write_ntuple)
@@ -385,7 +388,7 @@ void run_sdl()
     std::vector<SDL::Event<SDL::Acc>*> events;
     for (int s = 0; s < ana.streams; s++)
     {
-        SDL::Event<SDL::Acc> *event = new SDL::Event<SDL::Acc>(ana.verbose>=2, queue);
+        SDL::Event<SDL::Acc> *event = new SDL::Event<SDL::Acc>(ana.verbose>=2, queues[s]);
         events.push_back(event);
     }
     float timeForEventCreation = full_timer.RealTime()*1000;