This PR fixed the timing issues found by @GNiendorf. There were two issues that caused the observed slowdowns.

1. In the standalone version, the `sdl_<backend>` binary was creating a single queue and using it for all streams. This was effectively serializing the computation into a single stream.
2. By fixing the race conditions in #387, the loops ended up checking many more objects, slowing things down. I rewrote the logic of the loops a bit to bring the timing down to the original levels.

These are some timing comparisons on cgpu-1.
```
*** This PR ***
Total Timing Summary
Average time for map loading = 456.638 ms
Average time for input loading = 7443.24 ms
Average time for SDL::Event creation = 0.00727744 ms
   Evt    Hits       MD       LS      T3       T5       pLS       pT5      pT3      TC       Reset    Event     Short             Rate
   avg     23.2      1.4      0.4      1.8      1.7      0.4      1.3      0.6      1.3      0.1      32.2       8.6+/-  1.7      34.5   explicit_cache[s=1]
   avg      8.2      1.5      0.5      2.4      2.1      0.5      1.9      0.9      1.8      0.2      20.0      11.4+/-  2.6      11.4   explicit_cache[s=2]
   avg     11.4      1.9      0.9      3.8      3.4      0.6      3.3      1.5      2.9      0.4      30.0      18.1+/-  4.3       8.3   explicit_cache[s=4]
   avg     17.4      2.5      1.2      5.0      4.6      0.7      4.5      2.2      4.3      0.5      42.9      24.8+/-  5.3       7.6   explicit_cache[s=6]
   avg     24.5      3.2      1.6      6.4      6.0      0.9      5.5      2.8      5.6      0.8      57.2      31.8+/-  6.6       7.7   explicit_cache[s=8]

*** master ***
Total Timing Summary
Average time for map loading = 459.319 ms
Average time for input loading = 7439.52 ms
Average time for SDL::Event creation = 0.0067552 ms
   Evt    Hits       MD       LS      T3       T5       pLS       pT5      pT3      TC       Reset    Event     Short             Rate
   avg     27.1      1.4      0.4      1.8      1.7      0.4      0.6      0.5      6.1      0.1      40.0      12.5+/-  6.1      42.2   explicit_cache[s=1]
   avg     10.7      2.0      0.5      5.2      2.5      0.5      2.8      2.4      6.3      0.4      33.3      22.1+/-  8.0      17.9   explicit_cache[s=2]
   avg     21.3      4.0      1.1      8.6      5.5      1.1      7.1      5.9      8.2      0.8      63.5      41.2+/- 12.4      16.6   explicit_cache[s=4]
   avg     33.1      6.1      1.6     13.0      9.3      1.2     10.1     10.5     10.2      1.9      96.9      62.7+/- 16.6      16.8   explicit_cache[s=6]
   avg     46.7      7.5      1.4     15.4     12.5      1.9     16.2     15.0     11.7      2.6     130.7      82.1+/- 19.7      16.9   explicit_cache[s=8]

*** old commit before any of these issues (fc97a6b) ***
Total Timing Summary
Average time for map loading = 838.334 ms
Average time for input loading = 7260.38 ms
Average time for SDL::Event creation = 0.0427564 ms
   Evt    Hits       MD       LS      T3       T5       pLS       pT5      pT3      TC       Reset    Event     Short             Rate
   avg     26.5      1.4      0.4      1.8      1.5      0.6      0.8      0.6      1.4      0.1      35.2       8.0+/-  1.6      36.7   explicit_cache[s=1]
   avg      8.2      1.5      0.5      2.4      2.0      0.6      1.3      0.9      1.9      0.2      19.6      10.7+/-  2.4      10.7   explicit_cache[s=2]
   avg     11.3      1.9      0.8      3.8      3.2      0.7      2.5      1.6      2.9      0.4      29.2      17.1+/-  3.5       7.8   explicit_cache[s=4]
   avg     17.0      2.5      1.3      5.1      4.8      0.9      3.8      2.2      4.2      0.9      42.6      24.8+/-  6.1       7.6   explicit_cache[s=6]
   avg     24.4      3.2      1.6      6.6      6.3      1.0      4.8      2.9      5.4      1.1      57.4      32.0+/-  8.1       7.6   explicit_cache[s=8]
```