@ariostas 2024-04-29
As you can see, the standalone check now includes timing information. I ran some tests and it turned out to be fairly consistent. It should be reliable enough to let us know of significant slowdowns so that this sort of thing doesn't happen again.

The efficiency has some small variations and the fake rate and duplicate rate have small increases compared to master. I'm looking to see if it's possible to keep them as close to how they were before without increasing computing time or introducing race conditions.