@GNiendorf 2024-04-29
<img width="1411" alt="Screenshot 2024-04-29 at 3 00 19 PM" src="https://github.com/SegmentLinking/TrackLooper/assets/25272611/3700e97f-bfae-4420-a4c3-2a826535c5e1">

Timing looks normal again on 4555 as well, thank you for the fix!